+++ src/screen/HomeScreen/screen_Schedule.js
@@ -1,0 +1,18 @@
+import {
+    Image,
+    View,
+    Text,
+    StyleSheet,
+    Button,
+    TouchableOpacity,
+    ScrollView,
+  
+  } from 'react-native';
+
+export default function Schedule() {
+    return(
+        <>
+        <Text>Schedule</Text>
+        </>
+    )
+}
