+++ src/screen/HomeScreen/screen_Tuition.js
@@ -1,0 +1,18 @@
+import {
+    Image,
+    View,
+    Text,
+    StyleSheet,
+    Button,
+    TouchableOpacity,
+    ScrollView,
+  
+  } from 'react-native';
+
+export default function Tuition() {
+    return(
+        <>
+        <Text>Tuition</Text>
+        </>
+    )
+}
